requirements.txt
requests
pyyaml
pystray
Pillow
